# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Price
  description: 'Generate price'
  version: 1.0.0
tags:
  - name: Price
    description: Price
paths:
  /prices:
    get:
      tags:
        - Price
      summary: Returns information prices
      operationId: getPrice
      parameters:
        - in: query
          name: productId
          allowEmptyValue: false
          description: Product identifier.
          required: true
          schema:
            type: integer
        - in: query
          name: brandId
          allowEmptyValue: false
          description: Group identifier.
          required: true
          schema:
            type: integer
        - in: query
          name: applyDate
          allowEmptyValue: false
          description: date (yyyy-MM-dd-HH.mm.ss)
          required: true
          schema:
            type: string
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'

        400:
          description: Input parameters error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/PriceError'
              example:
                status: 400
                title: Bad Request
                detail: Date formatting error

        500:
          description: Unexpected problem server
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/PriceError'
              example:
                status: 500
                title: Internal server error
                detail: Unexpected server error
components:
  schemas:
    PriceResponse:
      description: Object search find result
      type: object
      properties:
        productId:
          description: Identifier product
          type: integer
        brandId:
          description: Identifier brand
          type: integer
        priceList:
          description: List of prices
          type: integer
        startDate:
          description: start date
          type: string
          format: date-time
          example: yyyy-MM-dd-HH.mm.ss
        endDate:
          description: end date
          type: string
          format: date-time
          example: yyyy-MM-dd-HH.mm.ss
        price:
          description: Final price applied
          type: number
          format: double
    PriceError:
      type: object
      properties:
        title:
          type: string
          description: Description of the problem type
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
        detail:
          type: string
          description: Detail of Problem